pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mergeguard"
version = "0.1.0"
description = "AI PR Code Review Agent powered by AWS Strands SDK and Amazon Bedrock"
requires-python = ">=3.12,<3.14"
dependencies = [
    "strands-agents>=0.1.0",
    "boto3>=1.34.0",
    "tree-sitter>=0.21.0",
    "tree-sitter-languages>=1.10.0",
    "unidiff>=0.7.5",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "networkx>=3.3",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "PyJWT[cryptography]>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-recording>=0.13.0",
    "pytest-cov>=5.0.0",
    "vcrpy>=6.0.1",
    "types-pyyaml>=6.0.12",
    "types-networkx>=3.3.0",
]

[project.scripts]
mergeguard = "mergeguard.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/mergeguard"]

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
warn_unused_ignores = false
# strict = true — re-enable once the Phase 0–5 typing debt is paid down.
# Subset of checks kept on; broad strictness too noisy for current codebase.
check_untyped_defs = true
warn_redundant_casts = true
warn_unreachable = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=mergeguard --cov-report=term-missing"

[tool.coverage.run]
source = ["src/mergeguard"]
